// src/app/enums/STATE.ts
export enum STATE {
    IDLE = "IDLE",
    MAIN_VIEW = "MAIN_VIEW",
    FIELD_EDIT = "FIELD_EDIT",
    FIELD_VIEW = "FIELD_VIEW",
    TEMPLATE_EDIT = "TEMPLATE_EDIT",
    TEMPLATE_VIEW = "TEMPLATE_VIEW",
    RECORD_EDIT = "RECORD_EDIT",
    RECORD_VIEW = "RECORD_VIEW",
    RECORD_CREATE = "RECORD_CREATE",
    SUPER_USER_PANEL_VIEW = "SUPER_USER_PANEL_VIEW",
    SUPER_USER_PANEL_EDIT = "SUPER_USER_PANEL_EDIT"
}